specify float32 type when importing images

+++ dataset.py
@@ -16,8 +16,8 @@ class BinaryDataset(Dataset):
     def __getitem__(self, index):
         img_path = os.path.join(self.image_dir, self.images[index])
         mask_path = os.path.join(self.mask_dir, self.images[index])
-        image = np.array(Image.open(img_path))
-        mask = np.array(Image.open(mask_path), dtype=np.float32) # open labelled image
+        image = np.array(Image.open(img_path)).astype(np.float32)
+        mask = np.array(Image.open(mask_path)).astype(np.float32) # open labelled image
 
         if self.transform is not None:
             augmentations = self.transform(image=image, mask=mask)
